Added effects.
Renamed the pages to index.html and sd_effects.css. Added text shadow to the h1. Added a background with ton of gradients and colors to the page. Added image affects to an image. Added a transform to that same image. Added a box shadow around the body. Added a border around the navigation. Added a top border to the footer.